requirements.txt
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
bs4==0.0.1
cftime==1.0.3.4
netCDF4==1.5.0.1
numpy==1.16.3
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
soupsieve==1.9.1
